Move ContentType class to PSR-4 locaiton
Pretty simple change to move the ContentType class into the Models folder - updates stack to use the new location.
